# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorEnvelope": {
        "properties": {
          "data": {
            "type": "null"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta",
          "error"
        ],
        "type": "object"
      },
...
      "CreateListingDto": {
        "properties": {
          "address": {
            "maxLength": 300,
            "type": "string"
          },
          "availability": {
            "items": {
              "$ref": "#/components/schemas/ListingAvailabilityDto"
            },
            "maxItems": 20,
            "type": "array"
          },
          "category": {
            "maxLength": 120,
            "type": "string"
          },
          "checkoutTime": {
            "type": "string"
          },
          "city": {
            "maxLength": 120,
            "type": "string"
          },
          "description": {
            "maxLength": 5000,
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "listingType": {
            "enum": [
              "private_room",
              "entire_home",
              "shared_room"
            ],
            "type": "string"
          },
          "localRecommendations": {
            "items": {
              "maxLength": 120,
              "type": "string"
            },
            "maxItems": 20,
            "type": "array"
          },
          "longitude": {
            "type": "number"
          },
          "neighborhoodPerks": {
            "items": {
              "maxLength": 120,
              "type": "string"
            },
            "maxItems": 20,
            "type": "array"
          },
          "placeId": {
            "maxLength": 255,
            "type": "string"
          },
          "priceCents": {
            "minimum": 0,
            "type": "number"
          },
          "priceUnit": {
            "enum": [
              "day",
              "night",
              "month"
            ],
            "type": "string"
          },
          "proximityTags": {
            "items": {
              "enum": [
                "near_hospitals",
                "public_transit"
              ],
              "type": "string"
            },
            "maxItems": 10,
            "type": "array"
          },
          "specialFeatures": {
            "items": {
              "enum": [
                "fully_furnished",
                "pet_friendly",
                "access_24_7"
              ],
              "type": "string"
            },
            "maxItems": 20,
            "type": "array"
          },
          "stayDurations": {
            "items": {
              "enum": [
                "short_term",
                "medium_term",
                "long_term"
              ],
              "type": "string"
            },
            "maxItems": 10,
            "type": "array"
          },
          "timeZone": {
            "maxLength": 64,
            "type": "string"
          },
          "title": {
            "maxLength": 140,
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "city",
          "timeZone",
          "priceCents",
          "priceUnit",
          "listingType"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "Supabase JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "metricsToken": {
        "description": "Dedicated operations metrics token; not a Supabase token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Stable backend-only REST contracts. All ordinary JSON responses use the MediCN data/meta/error envelope. Socket.IO is documented separately.",
    "title": "MediCN backend API",
    "version": "5.9C"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/listings": {
      "get": {
        "operationId": "ListingsController_searchListings",
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "nearbyHospital",
            "required": false,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "listingType",
            "required": false,
            "schema": {
              "enum": [
                "private_room",
                "entire_home",
                "shared_room"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stayDuration",
            "required": false,
            "schema": {
              "enum": [
                "short_term",
                "medium_term",
                "long_term"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minPrice",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "maxPrice",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "priceUnit",
            "required": false,
            "schema": {
              "enum": [
                "day",
                "night",
                "month"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "enum": [
                "newest",
                "price_asc",
                "price_desc"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "bounds",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 429 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 503 in the stable error envelope."
          }
        },
        "security": [],
        "tags": [
          "Listings"
        ]
      },
      "post": {
        "operationId": "ListingsController_createListing",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Listings"
        ]
      }
    },
    "/api/v1/listings/{id}": {
      "delete": {
        "operationId": "ListingsController_deleteListing",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Listings"
        ]
      },
      "get": {
        "operationId": "ListingsController_getListing",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Listings"
        ]
      },
      "patch": {
        "operationId": "ListingsController_updateListing",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Listings"
        ]
      }
    },
    "/api/v1/listings/{id}/availability": {
      "get": {
        "operationId": "ListingsController_listAvailability",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Listings"
        ]
      },
      "post": {
        "operationId": "ListingsController_createAvailability",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAvailabilityWindowDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Listings"
        ]
      }
    },
    "/api/v1/listings/{id}/availability/{windowId}": {
      "delete": {
        "operationId": "ListingsController_deleteAvailability",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "windowId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Listings"
        ]
      },
      "patch": {
        "operationId": "ListingsController_updateAvailability",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "windowId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvailabilityWindowDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 400 in the stable error envelope."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 401 in the stable error envelope."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 403 in the stable error envelope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 404 in the stable error envelope."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 413 in the stable error envelope."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "HTTP 500 in the stable error envelope."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Listings"
        ]
      }
    },
...
  },
  "servers": [],
...
}